Production-hardening 576: fix i64 overflow ICE/UB in date/time floor_mod near timestamp extremes
time.rs's shared floor_mod(a, b), used by both interp and vm, computed
a - floor_div(a, b) * b. The final result is always in [0, b) (b is always
a small constant: 86400, 400, 146097, 7), but the intermediate product
floor_div(a, b) * b can transiently overflow i64 when a is a raw Unix
timestamp spanning the full i64 range -- a perfectly valid Int with no
type-level restriction against extreme values. In a debug build this
tripped Rust's overflow panic, caught by main.rs's generic panic hook and
misreported as an "internal compiler error" -- crashing the interpreter/KVM
on an ordinary date_iso/year_of/weekday_of call with a large-magnitude
timestamp, not a genuine language bug.

cgen.rs's k_floor_mod mirrors the same shape in C, where signed integer
overflow is undefined behavior rather than a caught panic. Confirmed via
UBSan on the same repro; the non-sanitized native build silently produced
a value anyway, exactly the "looks fine under normal testing, only a
sanitizer catches it" pattern this campaign's native-runtime-audit vein
exists to find.

Fixed both by widening the multiply/subtract to a 128-bit intermediate
(i128 in Rust, __int128 in C) -- the final result is always in-bounds by
construction, so this fully eliminates the overflow with no behavior change
for any in-range input.

Found via a static+ASan audit that also confirmed AI-fun's native JSON
tool-call marshaling fully reuses the already-fixed k_json_parse/kjp_string
(no independent parser to audit) and that date/time's ISO parsing and
civil-calendar conversion arithmetic are otherwise sound.

cargo test: 926 passing (924 lib + 2 main), zero failures. interp==vm
byte-identical across all examples/*.kupl. Native sweep: 57 pass/0 diverge/
0 native-fail/8 skipped. Re-verified clean under a fresh ASan+UBSan build.

+++ src/time.rs
@@ -20,7 +20,14 @@ fn floor_div(a: i64, b: i64) -> i64 {
 }
 
 fn floor_mod(a: i64, b: i64) -> i64 {
-    a - floor_div(a, b) * b
+    // The final result is always in [0, b) (tiny for every real caller here --
+    // 86400, 400, 146097, 7...), but the INTERMEDIATE product `floor_div(a, b) * b`
+    // can transiently exceed i64 range when `a` spans the full i64 timestamp range
+    // (e.g. a timestamp near i64::MIN/MAX) even though `a` itself and the final
+    // subtracted result both fit -- a classic "intermediate overflow, final value in
+    // range" trap. Widen to i128 for the multiply/subtract so this never overflows;
+    // the final cast back to i64 is always in-bounds by construction.
+    (a as i128 - floor_div(a, b) as i128 * b as i128) as i64
 }
 
 /// (year, month 1..=12, day 1..=31) from a count of days since 1970-01-01.
